requirements.txt
rich>=13.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
PyYAML>=6.0
requests>=2.31.0
